pyproject.toml
[project]
name = "collectors"
version = "0.1.0"
description = "pokemon"
authors = [
    { name = "yole", email = "yole@hajduk.com" }
]
dependencies = [
    "Django==5.1.4",
    "python-dotenv==1.0.1",
    "psycopg2-binary==2.9.10",
    "django-admin-rangefilter==0.13.2"
]

[project.optional-dependencies]
dev = ["black==23.9.1", "pre-commit==3.5.0"]

[tool.setuptools]
packages = ["main", "user", "collector"]
